refactor(cx): move legacy class-ref parsing into compat
